requirements.txt
python-dateutil
python-decouple
psycopg2-binary
staticmap
sodapy
tweepy
